# contracts/gateway.openapi.json
{
  "components": {
    "schemas": {
      "Body_speech_to_text_v1_stt_post": {
        "properties": {
          "audio": {
            "contentMediaType": "application/octet-stream",
            "description": "Voice note: wav, m4a/aac, mp3, ogg/opus, flac, webm, 3gp",
            "title": "Audio",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "pattern": "^[a-z]{2,3}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO 639-1; omit to auto-detect",
            "title": "Language"
          }
        },
        "required": [
          "audio"
        ],
        "title": "Body_speech_to_text_v1_stt_post",
        "type": "object"
      },
      "EntryOut": {
        "properties": {
          "carbs": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carbs"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "day": {
            "format": "date",
            "title": "Day",
            "type": "string"
          },
          "deleted": {
            "default": false,
            "title": "Deleted",
            "type": "boolean"
          },
          "eaten_at": {
            "format": "date-time",
            "title": "Eaten At",
            "type": "string"
          },
          "fat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fat"
          },
          "food_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The database food this entry was matched to",
            "title": "Food Name"
          },
          "grams": {
            "title": "Grams",
            "type": "number"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kcal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when the food was not found (status=unmatched)",
            "title": "Kcal"
          },
          "meal_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meal Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "per100": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Per100"
              },
              {
                "type": "null"
              }
            ],
            "description": "Lets the client rescale offline: value = per100 * grams / 100"
          },
          "position": {
            "description": "Order within one message; sort the diary by (eaten_at, position)",
            "title": "Position",
            "type": "integer"
          },
          "protein": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protein"
          },
          "source": {
            "enum": [
              "text",
              "voice",
              "photo"
            ],
            "title": "Source",
            "type": "string"
          },
          "status": {
            "enum": [
              "ok",
              "uncertain",
              "unmatched"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "day",
          "eaten_at",
          "position",
          "name",
          "grams",
          "status",
          "confidence",
          "source",
          "updated_at"
        ],
        "title": "EntryOut",
        "type": "object"
      },
      "EntryPatchIn": {
        "properties": {
          "grams": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 5000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grams"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "EntryPatchIn",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "GoalIn": {
        "properties": {
          "calorie_goal": {
            "description": "kcal per day; values below the safety minimum are refused (goal_too_low)",
            "title": "Calorie Goal",
            "type": "integer"
          }
        },
        "required": [
          "calorie_goal"
        ],
        "title": "GoalIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeOut": {
        "properties": {
          "calorie_goal": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calorie Goal"
          },
          "email": {
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "MeOut",
        "type": "object"
      },
      "MessageIn": {
        "properties": {
          "client_id": {
            "description": "Generated by the app; retrying with the same id never applies a message twice",
            "format": "uuid",
            "title": "Client Id",
            "type": "string"
          },
          "day": {
            "description": "The user's local date for this entry",
            "format": "date",
            "title": "Day",
            "type": "string"
          },
          "eaten_at": {
            "description": "Local time with UTC offset, e.g. 2026-09-30T08:15:00+03:00",
            "format": "date-time",
            "title": "Eaten At",
            "type": "string"
          },
          "image_base64": {
            "anyOf": [
              {
                "maxLength": 6000000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "JPEG/PNG/WebP bytes, base64",
            "title": "Image Base64"
          },
          "image_mime": {
            "default": "image/jpeg",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ],
            "title": "Image Mime",
            "type": "string"
          },
          "pending_question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PendingQuestionIn"
              },
              {
                "type": "null"
              }
            ],
            "description": "Send the question the app last showed; the text is then read as its answer"
          },
          "source": {
            "default": "text",
            "enum": [
              "text",
              "voice",
              "photo"
            ],
            "title": "Source",
            "type": "string"
          },
          "text": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "required": [
          "client_id",
          "day",
          "eaten_at"
        ],
        "title": "MessageIn",
        "type": "object"
      },
      "MessageOut": {
        "properties": {
          "clarify_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "At most one short question to show in the chat",
            "title": "Clarify Question"
          },
          "clarify_target_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Entry the question is about, if any",
            "title": "Clarify Target Id"
          },
          "entries": {
            "description": "Entries created or changed by this message, in order",
            "items": {
              "$ref": "#/components/schemas/EntryOut"
            },
            "title": "Entries",
            "type": "array"
          },
          "removed_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Removed Ids",
            "type": "array"
          }
        },
        "required": [
          "entries"
        ],
        "title": "MessageOut",
        "type": "object"
      },
...
      "SttOut": {
        "properties": {
          "duration_s": {
            "title": "Duration S",
            "type": "number"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "language",
          "duration_s"
        ],
        "title": "SttOut",
        "type": "object"
      },
      "SyncOut": {
        "properties": {
          "calorie_goal": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calorie Goal"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/EntryOut"
            },
            "title": "Entries",
            "type": "array"
          },
          "has_more": {
            "description": "True when a table hit the page size; call again right away with next_since",
            "title": "Has More",
            "type": "boolean"
          },
          "next_since": {
            "description": "Pass as `since` on the next call. Re-delivered rows are harmless: upsert by id.",
            "format": "date-time",
            "title": "Next Since",
            "type": "string"
          },
          "server_time": {
            "format": "date-time",
            "title": "Server Time",
            "type": "string"
          },
          "weights": {
            "items": {
              "$ref": "#/components/schemas/WeightOut"
            },
            "title": "Weights",
            "type": "array"
          }
        },
        "required": [
          "server_time",
          "next_since",
          "has_more",
          "entries",
          "weights"
        ],
        "title": "SyncOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Token from POST /v1/auth/verify",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "DietApp gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/entries/{entry_id}": {
      "delete": {
        "operationId": "delete_entry_v1_entries__entry_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Entry",
        "tags": [
          "diary"
        ]
      },
      "patch": {
        "description": "Edit a row in place. Changing only grams is pure arithmetic; changing the name looks the food up again.",
        "operationId": "patch_entry_v1_entries__entry_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryPatchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Entry",
        "tags": [
          "diary"
        ]
      }
    },
    "/v1/me": {
      "get": {
        "operationId": "me_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "diary"
        ]
      }
    },
    "/v1/me/goal": {
      "put": {
        "description": "Set the daily calorie goal. Goals under the safety minimum are refused, never adjusted silently.",
        "operationId": "set_goal_v1_me_goal_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Goal",
        "tags": [
          "diary"
        ]
      }
    },
    "/v1/messages": {
      "post": {
        "description": "Turn a text and/or photo message into diary entries (or changes to today's entries).\n\nIdempotent on `client_id`: a retry returns the stored answer. 5xx/504 mean \"try again later\";\nthe same client_id is safe to reuse.",
        "operationId": "post_message_v1_messages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Post Message",
        "tags": [
          "diary"
        ]
      }
    },
    "/v1/stt": {
      "post": {
        "description": "Fallback for devices without on-device recognition. The text goes back to the app, which sends it as a message.",
        "operationId": "speech_to_text_v1_stt_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_speech_to_text_v1_stt_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SttOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Speech To Text",
        "tags": [
          "diary"
        ]
      }
    },
    "/v1/sync": {
      "get": {
        "description": "Everything changed at or after `since`, including deletions (deleted=true). Upsert by id on the client.",
        "operationId": "sync_v1_sync_get",
        "parameters": [
          {
            "description": "`next_since` from the previous call; omit for everything",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`next_since` from the previous call; omit for everything",
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sync",
        "tags": [
          "diary"
        ]
      }
    },
    "/v1/weights/{weight_id}": {
      "delete": {
        "operationId": "delete_weight_v1_weights__weight_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "weight_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Weight Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Weight",
        "tags": [
          "diary"
        ]
      },
      "put": {
        "description": "Create or replace a weight by its client-generated id (safe to repeat).",
        "operationId": "put_weight_v1_weights__weight_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "weight_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Weight Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeightIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Put Weight",
        "tags": [
          "diary"
        ]
      }
    }
  }
}